@@ -1,14 +1,14 @@
 {
   "name": "noblox.js",
-  "version": "4.15.1",
+  "version": "6.0.2",
   "description": "A Node.js API wrapper for Roblox.",
   "main": "lib/index.js",
   "types": "typings/index.d.ts",
   "scripts": {
     "docs": "jsdoc -c jsDocsConfig.json -r -t ./node_modules/better-docs",
-    "lint": "eslint lib/",
+    "lint": "standard lib/",
     "test": "jest",
-    "postinstall": "node postinstall.js"
+    "postinstall": "node postinstall.mjs"
   },
   "repository": {
     "type": "git",
@@ -20,6 +20,7 @@
     "clans"
   ],
   "dependencies": {
+<<<<<<< HEAD
     "@microsoft/signalr": "^8.0.0",
     "chalk": "^4.0.0",
     "cheerio": "^1.0.0-rc.10",
@@ -27,6 +28,14 @@
     "figlet": "^1.4.0",
     "request": "^2.81.0",
     "request-promise": "^4.2.4"
+=======
+    "@microsoft/signalr": "^8.0.7",
+    "chalk": "^5.3.0",
+    "cheerio": "^1.0.0-rc.10",
+    "entities": "^5.0.0",
+    "figlet": "^1.7.0",
+    "postman-request": "^2.88.1-postman.34"
+>>>>>>> upstream/master
   },
   "license": "MIT",
   "bugs": {
@@ -40,6 +49,7 @@
   ],
   "homepage": "https://github.com/noblox/noblox.js",
   "devDependencies": {
+<<<<<<< HEAD
     "@auto-it/conventional-commits": "^10.30.0",
     "auto": "^10.30.0",
     "better-docs": "2.7.2",
@@ -56,6 +66,17 @@
     "JSONStream": "^1.3.1",
     "progress": "^2.0.0",
     "standard": "^17.0.0"
+=======
+    "@auto-it/conventional-commits": "^11.2.0",
+    "JSONStream": "^1.3.1",
+    "auto": "^11.2.0",
+    "better-docs": "2.7.3",
+    "dotenv": "16.4.5",
+    "jest": "^29.7.0",
+    "jest-extended": "^4.0.2",
+    "jsdoc": "4.0.3",
+    "standard": "^17.1.0"
+>>>>>>> upstream/master
   },
   "jest": {
     "setupFiles": [
@@ -83,6 +104,7 @@
     "onlyPublishWithReleaseLabel": true
   },
   "engines": {
-    "node": ">=14.14.0"
-  }
+    "node": ">=18.18"
+  },
+  "packageManager": "yarn@1.22.19+sha1.4ba7fc5c6e704fce2066ecbfb0b0d8976fe62447"
 }
