@@ -1,5 +1,9 @@
 // Dependencies
+<<<<<<< HEAD
 const { HubConnectionBuilder, HttpTransportType } = require('@microsoft/signalr')
+=======
+const signalR = require('@microsoft/signalr')
+>>>>>>> upstream/master
 const events = require('events')
 
 // Includes
@@ -37,6 +41,7 @@ exports.func = function (args) {
       args.jar = { session: args.jar }
     }
     const session = getSession({ jar: args.jar })
+<<<<<<< HEAD
     const headers = {
       Cookie: '.ROBLOSECURITY=' + session + ';'
       // Add other headers if needed
@@ -77,6 +82,27 @@ exports.func = function (args) {
     client.serviceHandlers = client.serviceHandlers || {}
     client.serviceHandlers.connectFailed = function (err) {
       console.error('Connection failed:', err)
+=======
+    let userNotificationConnection = null
+
+    userNotificationConnection = new signalR.HubConnectionBuilder()
+      .withUrl('https://realtime-signalr.roblox.com/userhub', {
+        transport: signalR.HttpTransportType.WebSockets,
+        skipNegotiation: true,
+        headers: {
+          Cookie: '.ROBLOSECURITY=' + session + ';'
+        }
+      })
+      .build()
+
+    userNotificationConnection.on('notification', function (name, message) {
+      notifications.emit('data', name, JSON.parse(message))
+    })
+
+    notifications.on('close', userNotificationConnection.stop)
+
+    userNotificationConnection.disconnected = function (err) {
+>>>>>>> upstream/master
       notifications.emit('error', new Error('Connection failed: ' + err.message))
       if (retries !== -1) {
         if (retries > max) {
@@ -86,18 +112,31 @@ exports.func = function (args) {
         }
       }
     }
+<<<<<<< HEAD
     client.serviceHandlers.onerror = function (err) {
       console.error('Connection error:', err)
+=======
+
+    userNotificationConnection.error = function (err) {
+>>>>>>> upstream/master
       notifications.emit('error', err)
     }
-    client.serviceHandlers.connected = function (connection) {
+
+    userNotificationConnection.connected = function (connection) {
       notifications.emit('connect', connection)
     }
+<<<<<<< HEAD
     client.serviceHandlers.reconnecting = function () {
+=======
+
+    userNotificationConnection.reconnecting = function () {
+      setTimeout(connect, 5000, 0)
+>>>>>>> upstream/master
       notifications.emit('error', new Error('Lost connection, reconnecting'))
       return true // Abort reconnection
     }
 
+<<<<<<< HEAD
     try {
       await client.start()
       return client
@@ -105,6 +144,9 @@ exports.func = function (args) {
       console.error('Error while starting connection:', err)
       throw err
     }
+=======
+    userNotificationConnection.start()
+>>>>>>> upstream/master
   }
 
   connect(-1)
