@@ -1,5 +1,5 @@
 const options = require('../options.js')
-const getCurrentUser = require('../util/getCurrentUser.js').func
+const getAuthenticatedUser = require('../util/getAuthenticatedUser.js').func
 
 exports.required = ['cookie']
 exports.optional = ['validate']
@@ -11,7 +11,7 @@ exports.optional = ['validate']
  * @alias setCookie
  * @param {string} cookie - The cookie to sign in with.
  * @param {boolean=} [validate=true] - Whether to validate the cookie or not.
- * @returns {Promise<LoggedInUserData>}
+ * @returns {Promise<AuthenticatedUserData>}
  * @example const noblox = require("noblox.js")
  * noblox.setCookie("cookie").then(function() {
  *   //your code here
@@ -28,7 +28,7 @@ exports.func = async function (args) {
     return false
   }
   try {
-    const res = await getCurrentUser({ jar: { session: args.cookie } })
+    const res = await getAuthenticatedUser({ jar: { session: args.cookie } })
     options.jar.session = args.cookie
     return res
   } catch (error) {
