@@ -1,6 +1,5 @@
 const http = require('../util/http.js').func
 const getGeneralToken = require('../util/getGeneralToken.js').func
-const checkProductName = require('./checkDeveloperProductName.js').func
 
 exports.required = ['universeId', 'name', 'priceInRobux']
 exports.optional = ['description', 'jar']
@@ -21,37 +20,26 @@ exports.optional = ['description', 'jar']
 **/
 
 const nextFunction = (jar, token, universeId, name, priceInRobux, description) => {
-  return checkProductName({
-    universeId,
-    productName: name
-  }).then((res) => {
-    if (res.Success && res.Message === 'Name available') {
-      return http({
-        url: '//apis.roblox.com/developer-products/v1/universes/' + universeId + '/developerproducts?name=' + name + '&description=' + description + '&priceInRobux=' + priceInRobux,
-        options: {
-          method: 'POST',
-          jar,
-          headers: {
-            'X-CSRF-TOKEN': token
-          },
-          resolveWithFullResponse: true
-        }
-      }).then((res) => {
-        console.log(res)
-        if (res.statusCode === 200) {
-          return {
-            universeId,
-            name,
-            priceInRobux,
-            description,
-            productId: typeof res.body === 'object' ? res.body.id : JSON.parse(res.body).id
-          }
-        } else {
-          throw new Error('Create product failed, ' + res.statusCode + ' ' + res.statusMessage + '')
-        }
-      })
-    } else {
-      throw new Error('Product with this name already exists')
+  description = description || ''
+  return http({
+    url: '//apis.roblox.com/developer-products/v1/universes/' + universeId + '/developerproducts?name=' + name + '&description=' + description + '&priceInRobux=' + priceInRobux,
+    options: {
+      method: 'POST',
+      jar,
+      headers: {
+        'X-CSRF-TOKEN': token
+      },
+      resolveWithFullResponse: true
+    }
+  }).then(function (res) {
+    try {
+      const json = JSON.parse(res.body)
+      if (res.statusCode === 200) {
+        return json
+      }
+      throw new Error(json)
+    } catch (err) {
+      throw new Error(res.body)
     }
   })
 }
