@@ -19,7 +19,7 @@ exports.optional = ['pageNumber', 'pageSize', 'messageTab', 'jar']
 **/
 
 // Define
-function getMessages (jar, token, pageNumber, pageSize, messageTab) {
+function getMessages (jar, pageNumber, pageSize, messageTab) {
   return new Promise((resolve, reject) => {
     const httpOpt = {
       url: `//privatemessages.roblox.com/v1/messages?pageNumber=${pageNumber}&pageSize=${pageSize}&messageTab=${messageTab}`,
