@@ -34,19 +34,20 @@ exports.func = function (args) {
   const onMessage = new events.EventEmitter()
   let waitingForRequest = false
   let latest
-  getMessages({ jar, messageTab: 'Inbox', pageNumber: 0 })
+  getMessages({ jar, messageTab: 'Inbox', pageNumber: 0, pageSize: 1 })
     .then(function (initial) {
       latest = initial.collection[0] ? initial.collection[0].id : 0
       const notifications = onNotification({ jar })
       notifications.on('data', function (name, message) {
-        if (name === 'NotificationStream' && message.Type === 'NewNotification') {
+        if (name === 'MessageNotification' && message.Type === 'Created') {
           if (waitingForRequest) {
             waitingForRequest = false
           } else {
             getMessages({
               jar,
               messageTab: 'Inbox',
-              pageNumber: 0
+              pageNumber: 0,
+              pageSize: 1
             })
               .then(function (inbox) {
                 const messages = inbox.collection
