@@ -1,7 +1,8 @@
 // Includes
-const http = require('./http.js').func
+const settings = require('../../settings.json')
 
-// Args
+const noblox = require('../index.js')
+
 exports.optional = ['option', 'jar']
 
 // Docs
@@ -12,6 +13,7 @@ exports.optional = ['option', 'jar']
  * @param {string=} option - A specific option to return.
  * @returns {LoggedInUserData}
  * @example const noblox = require("noblox.js")
+ * @deprecated getCurrentUser() is deprecated; see getAuthenticatedUser(), getPremium(), getThumbnails(), getUserFunds() instead | August 27, 2024 - https://devforum.roblox.com/t/official-list-of-deprecated-web-endpoints/62889/66
  * // Login using your cookie.
  * const user = await noblox.getCurrentUser()
 **/
@@ -20,23 +22,35 @@ exports.optional = ['option', 'jar']
 exports.func = async function (args) {
   const jar = args.jar
   const option = args.option
-  const httpOpt = {
-    url: '//www.roblox.com/mobileapi/userinfo',
-    options: {
-      resolveWithFullResponse: true,
-      method: 'GET',
-      followRedirect: false,
-      jar
-    }
+  if (settings.show_deprecation_warnings) {
+    console.warn('[DEPRECATED]: getCurrentUser() is deprecated by Roblox; use getAuthenticatedUser(), getPremium(), getThumbnails(), or getUserFunds() instead!')
+    console.warn(' > Opt out of these warnings using noblox.setOptions({ show_deprecation_warnings: false })')
   }
-  const res = await http(httpOpt)
-  if (res.statusCode !== 200) {
-    throw new Error('You are not logged in.')
+
+  const currentUser = await noblox.getAuthenticatedUser(jar)
+  const [premiumStatus, thumbnailResponse, robuxBalance] = await Promise.all(
+    [
+      noblox.getPremium(currentUser.id, jar),
+      noblox.getPlayerThumbnail(currentUser.id, '352x352', 'png', false, 'Body', jar),
+      noblox.getUserFunds(currentUser.id, jar)
+    ]
+  )
+
+  const json = {
+    UserID: currentUser.id,
+    UserName: currentUser.name,
+    RobuxBalance: robuxBalance,
+    ThumbnailUrl: thumbnailResponse[0].imageUrl,
+    IsAnyBuildersClubMember: false,
+    IsPremium: premiumStatus,
+    DEPRECATION_WARNING: '[DEPRECATED]: noblox.getCurrentUser() is deprecated; use getAuthenticatedUser(), getPremium(), getThumbnails(), or getUserFunds() instead!'
   }
-  const json = JSON.parse(res.body)
+
   if (!option) {
     return json
   }
+
+  // Support queried rgequests `getCurrentUser('UserID') -> only UserID`
   const searchKey = Object.keys(json).filter((key) => {
     return option.toLowerCase() === key.toLowerCase()
   })[0]
