@@ -1,5 +1,6 @@
 // Dependencies
-const request = require('request-promise')
+const util = require('util')
+const request = util.promisify(require('postman-request'))
 
 // Includes
 const settings = require('../../settings.json')
