@@ -6,7 +6,7 @@ const settings = require('../../settings.json')
  * altering the settings.json file. Objects passed to this function should match the format of the settings.json file.
  * Unknown keys, or malformed options will be rejected with an error.
  * @category Utility
- * @param {NobloxOptions} newOptions - The new options to set, structured as per [settings.json](https://github.com/noblox/noblox.js/blob/master/settings.json)
+ * @param {Partial<NobloxOptions>} newOptions - The new options to set, structured as per [settings.json](https://github.com/noblox/noblox.js/blob/master/settings.json)
  * @returns void
  * @see [settings.json](https://github.com/noblox/noblox.js/blob/master/settings.json) - default package settings
  * @example const noblox = require("noblox.js")
